# swytchcode/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SupportIQ Swytchcode Governed API",
    "version": "1.0.0",
    "description": "Swytchcode API manifest for e-commerce customer support tools."
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "SupportIQ Local FastAPI Backend Server"
    }
  ],
  "paths": {
    "/tools/order-status": {
      "post": {
        "summary": "Check Order Status",
        "operationId": "get_order_status",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order_id": { "type": "string", "example": "48291" }
                },
                "required": ["order_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "order_id": { "type": "string" },
                    "status": { "type": "string" },
                    "carrier": { "type": "string" },
                    "tracking_number": { "type": "string" },
                    "estimated_delivery": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tools/request-refund": {
      "post": {
        "summary": "Request Refund",
        "operationId": "request_refund",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order_id": { "type": "string" },
                  "amount": { "type": "number" },
                  "reason": { "type": "string" }
                },
                "required": ["order_id", "amount", "reason"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refund processed or queued for approval"
          }
        }
      }
    },
    "/tools/create-ticket": {
      "post": {
        "summary": "Create Support Ticket",
        "operationId": "create_support_ticket",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order_id": { "type": "string" },
                  "issue": { "type": "string" },
                  "priority": { "type": "string" }
                },
                "required": ["issue", "priority"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket created successfully"
          }
        }
      }
    }
  }
}